+++ app/cli/cli.ts
@@ -9,13 +9,11 @@
  * @license: MIT License
  *
  */
-import m from "@app/functions/module";
-import translate from "@translations/translate";
-import logger from "@app/utils/logger";
+import m from "@app/functions/module.js";
+import translate from "@translations/translate.js";
+import logger from "@app/utils/logger.js";
 
-(async () => {
-	const label = translate("hello", { name: "World" }); // This show "Hello World"! Is a literal template string from en.json
-	const { app } = await m({ text: label });
+const label = translate("hello", { name: "Patryk" }); // This show "Hello Patryk"! Is a literal template string from en.json
+const { app } = await m({ text: label });
 
-	logger.info(app());
-})();
+logger.info(app());