+++ app/utils/logger.ts
@@ -8,8 +8,9 @@
  * @license: MIT License
  *
  */
-import Logger, { LoggerOptions } from "@ptkdev/logger";
-import config from "@configs/config";
+import type { LoggerOptions } from "@ptkdev/logger";
+import Logger from "@ptkdev/logger";
+import config from "@configs/config.js";
 
 const options: LoggerOptions = config.logger as unknown as LoggerOptions; // typescript issue https://github.com/microsoft/TypeScript/issues/26552#issuecomment-484124880
 const logger = new Logger(options);